refactor(admin): extract block-form leaf modules (field-helpers, array-limits, file-accept)
Verbatim moves into src/routes/admin/client/block-form/: the six
cross-cutting pure helpers plus the shared imagePickerIconSvg
(field-helpers.ts), the array min/max boundary logic (array-limits.ts),
and the effectiveAccept computation with its module-level allowlist
cache (file-accept.ts). block-form.ts keeps re-exporting
checkArrayLimitReached and ArrayLimitInfo, so its compiled surface is
unchanged. coverage.mjs excludes the new dist subdirectory for the same
reason it excludes block-form.js: browser-only code that node:test
cannot drive.

Part of #38, slice 2/6.
